+++ rust/shekyl-crypto-pq/fuzz/fuzz_targets/fuzz_construct_output.rs
@@ -53,7 +53,7 @@ fuzz_target!(|data: &[u8]| {
         od.amount_tag,
         &od.enc_label,
         od.label_tag,
-        od.view_tag_x25519,
+        od.view_tag_prefilter,
         &spend_key,
         output_index,
     );
@@ -74,7 +74,7 @@ fuzz_target!(|data: &[u8]| {
             od.amount_tag,
             &od.enc_label,
             od.label_tag,
-            od.view_tag_x25519,
+            od.view_tag_prefilter,
             &spend_key,
             output_index,
         );
@@ -91,7 +91,7 @@ fuzz_target!(|data: &[u8]| {
         od.amount_tag.wrapping_add(1),
         &od.enc_label,
         od.label_tag,
-        od.view_tag_x25519,
+        od.view_tag_prefilter,
         &spend_key,
         output_index,
     );